*: move all implementation-specific packages under internal directory
Noone needs them, they are used only for tests and as examples.
