+++ internal/crypto/hash.go
