+++ internal/crypto/hash_test.go
