+++ internal/payload/compact.go
