+++ internal/payload/consensus_message.go
@@ -5,7 +5,7 @@ import (
 	"encoding/gob"
 
 	"github.com/nspcc-dev/dbft"
-	"github.com/nspcc-dev/dbft/crypto"
+	"github.com/nspcc-dev/dbft/internal/crypto"
 	"github.com/pkg/errors"
 )
