+++ internal/payload/helpers.go
