+++ internal/payload/recovery_message.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 
 	"github.com/nspcc-dev/dbft"
-	"github.com/nspcc-dev/dbft/crypto"
+	"github.com/nspcc-dev/dbft/internal/crypto"
 )
 
 type (
